Fixes argument handling in shell scripts
Addresses an issue in shell scripts where arguments were not being properly handled, potentially leading to incorrect program execution.

This commit updates the argument passing mechanism in the scripts to correctly handle multiple arguments, especially those containing spaces or special characters.
The fix ensures that arguments passed to the profiled binaries are correctly interpreted.

Also, fix load_ulong to be of correct type.

+++ src/client/protocol.h
@@ -11,7 +11,7 @@
 
 static inline uint32_t load_ulong(const void *p)
 {
-	unsigned long v;
+	uint32_t v;
 	memcpy(&v, p, sizeof v);
 	return v;
 }
